Refine jaxb-api dynamic access tests
